requirements.txt
Package             Version
------------------- ------------
asgiref             3.7.2
Django              5.0
djangorestframework 3.14.0
Pillow              10.1.0
pip                 22.3.1
psycopg2            2.9.9
python-dotenv       1.0.0
pytz                2023.3.post1
setuptools          65.5.0
sqlparse            0.4.4
tzdata              2023.3
